+++ fe2o3-amqp/src/link/receiver_link.rs
@@ -6,7 +6,7 @@ use serde_amqp::format_code::EncodingCodes;
 
 use crate::{
     endpoint::LinkExt,
-    util::{is_consecutive, AsByteIterator, IntoReader},
+    util::{is_consecutive, AsByteIterator, IntoReader, Sealed},
 };
 
 use super::{delivery::DeliveryInfo, *};
@@ -154,12 +154,11 @@ where
             .ok_or(Self::TransferError::DeliveryTagIsNone)?;
         let message_format = transfer.message_format;
 
-        let (message, mode) = if settled_by_sender {
+        let (result, mode) = if settled_by_sender {
             // If the message is pre-settled, there is no need to
             // add to the unsettled map and no need to reply to the Sender
-            let message = T::decode_into_message(payload.into_reader())
-                .map_err(|_| Self::TransferError::MessageDecodeError)?;
-            (message, None)
+            let result = T::decode_into_message(payload.into_reader());
+            (result, None)
         } else {
             // If the message is being sent settled by the sender, the value of this
             // field is ignored.
@@ -177,8 +176,7 @@ where
                 None => None,
             };
 
-            let message = T::decode_into_message(payload.into_reader())
-                .map_err(|_| Self::TransferError::MessageDecodeError)?;
+            let result = T::decode_into_message(payload.into_reader());
 
             let state = DeliveryState::Received(Received {
                 section_number, // What is section number?
@@ -201,7 +199,20 @@ where
                     .get_or_insert(OrderedMap::new())
                     .insert(delivery_tag.clone(), Some(state));
             }
-            (message, mode)
+            (result, mode)
+        };
+
+        let message = match result {
+            Ok(message) => message,
+            Err(source) => {
+                let info = DeliveryInfo {
+                    delivery_id,
+                    delivery_tag,
+                    rcv_settle_mode: mode,
+                    _sealed: Sealed {},
+                };
+                return Err(ReceiverTransferError::MessageDecodeError{info, source});
+            }
         };
 
         let link_output_handle = self